+++ PulseEffects/application.py
@@ -82,9 +82,6 @@ class Application(Gtk.Application):
             self.sie.post_messages(False)
             self.soe.post_messages(False)
 
-            self.sie.switch_on_all_apps = True
-            self.soe.switch_on_all_apps = True
-
             self.pm.find_sink_inputs()
             self.pm.find_source_outputs()