+++ CHANGELOG.md
@@ -2,10 +2,9 @@
 
 ## [Unreleased]
 ### Fixed
-- Do not force the "switch all apps" setting to true when running in service
+- Do not force the "switch all apps" setting to true when running in service. 
 - Fixed frequency value in band label being slightly different from the one
 in the frequency menu because of rounding problems
-mode
 
 ## [3.1.6]
 ### Added